Move function AppController.bindEventsAfterBibleTextLoaded to verseListController
